src/stress: add a generic stress test
The new source "stress.c" uses macros to be called with various types and
modes. It supports 32/64 bits, longs and blk, in unique or dup modes. It
is pretty similar to the stress suite in that it will pre-allocate many
nodes, and will randomly pick one of them, and either try to delete it if
it's present, or insert it with a random key if not present. It purposely
creates duplicates, allowing both the dup and unique code to test them.

It contains a few consistency checks (just the bare minimum to allow the
code to remain portable across types).

It shouldn't be too hard to port to ebtree for comparisons, and could
support strings and memory blocks using limited modifications.