src/opstime: add a new test to measure the per-operation time
This tool differs from the stress tester in that it measures the
average per-operation cost across a full tree of a given size for
various key series (including stdin). It is losely based on the
previous tool. For now it is a bit limited, mostly focusing on u32,
u64 and strings. It aims at implementing a hash table as well though
for now the buckets are just replaced with a counter so their
calculation cost is zero and they're evenly distributed. It supports
generation of series of values including full-scale random, random
increments, fixed increments, worst case, and stdin. The cost can be
calculated using averages, mean value or avg between half of the mean
values. Some work remains to be done.