import: rb: import an rbtree implementation at commit 8667831
This imports rbtree.c and rbtree.h from https://github.com/pvachon/rbtree
which appears to be a reasonably good implementation that is performant
and easy enough to use. The commit at import was 8667831 ("Merge pull
request #1 from masatake/patch-1"). It was stored into directory "rb"
and integrated into the main makefile to produce librbtree.a. The
CFLAGS had to be changed to support declaration after statement.