scripts: add scripts to run tests and plot results
The scripts iterates over multiple data types and hash sizes, and also
consumes some input files (which are not provided), to produce output
data. The gnuplot script consumes these data to produce graphs. It just
requires to set the data type in the TEST variable and have INPUT_FILE
point to the data file.